fix: 🐛 s 指定服务的帮助信息
